routers can recieve and make the message
one error of note is that the destination info doesn't show, and that
the ip address is set to 0.0.0.0.
I think this is a similar error as I've had, where when trying to access
info in the singleton class, it's not set to anything but the default.
for example, master.getNumberOfClients() results in number of clients
being 0, which should be false.